fix: Don't show opac note if library's 'showOpacNote' is disabled
Previously, when library's 'showOpacNote' was disabled, the Koha user's
opac note was still being displayed. This is because it's getting added
to _web_note regardless of the setting's value.
This patch removes the opacnote from Koha's getILSMessages, and adds the
setting check to when opacnote should be added to loadContactInformation
or not.