DIS-98 refactor: extract Privacy Settings
MyAccount/MyPrivacySettings is already referred to
as Privacy Settings in the patron preferences menu.

Extract the Privacy Settings templates and logic from
Cookie Consent related files, so that said
templates and logic can be used not only for
cookie consent-related features, but also for other
types of user consents, as fetched from Koha where
appropriate.