DIS-98 feat: edit consents given
The following applies only to patrons of libraries
which:

- use Koha as their ILS
- have koha-plugin-newsletter-consent plugin
enabled

Note that the plugin itself requires a minimum
Koha version of 23.11.00.000.

In MyAccount/MyPrivacySettings, patrons can:

- grant consent where they previously had not
- rescind consent they had previously given

Before testing, ensure that you have uploaded and
enabled the koha-plugin-newsletter-consent plugin
in your Koha instance. This was written against
version 1.0.2.

Also ensure that the PrivacyPolicyConsent System
setting in Koha is enabled.

Test plan:

1. Login as a patron
2. Navigate to MyAccount > Privacy Settings
3. Notice the 'Newsletter' section
5. Toggle the consent on, save, reload
6. Toggle the consent off, save, reload
7. Notice that your consent choice is saved each
time