DIS-98 feat: self registration privacy section
The following applies only to patrons of libraries
which:

- use Koha as their ILS
- have koha-plugin-newsletter-consent plugin
enabled

Note that the plugin itself requires a minimum
Koha version of 23.11.00.000.

Adds a 'Privacy' section under which Koha consent
types can be displayed, so that new patrons may
chose to give or decline giving their consent to
these (eg. 'Newsletter') as they register.

Before testing ensure that:

- Self Registration via ILS in Aspen is enabled
- your Koha instance has the
koha-plugin-newsletter-consent plugin enabled
- your Koha instance as the PrivacyPolicyConsent
system setting enabled

Test plan:

- on the Aspen home page, select 'SIGN IN'
- select 'Register for a new Library Card.'
- notice the 'Privacy' section
- notice the 'Newsletter' subsection
- either tick the checkbox within or not
- fill in any required fields
- submit the registration form
- login to Aspen as the patron you have created
(note: you may need to log in to the Koha staff
interface to assign a library card number to this
new patron)
- Navigate to Your Account > Privacy Settings
- notice that the Newsletter consent toggle
is set to 'ON' if you ticked the matching
consent checkbox when registering, and set to
'OFF' if you did not.