DIS-221 fix: run EBSCO EDS and host search
When attempting to run a search in 'Articles and
Databases' using the EBSCO EDS or host integration, the
attempt to display the results fails, and the
following error message is displayed instead:

SearchObject_EbscoEdsSearcher::processSearch():
Return value must be of type AspenError|array|null
, stdClass returned

For EBSCO host, the same error appear, but the
data type mentioned is SimpleXMLElement.

This bug was introduced in 25.01.00.

This patch adds stdClass to the list of valid
return types, restoring functionality to the
EBSCO EDS and host integration.

note: valid EBSCO EDS and host credentials are
required

To replicate:

- checkout to a 25.01.00-based branch
- enable the EBSCO EDS / host module
- create an EBSCO EDS / host setting
- run a search in 'Article and Databases'
- notice that EBSCO EDS / host search
results do not display
- notice the error message

Test plan:

- apply this patch
- run a search in 'Article and Databases'
- notice that EBSCO EDS / host search results
display normally