DIS-138 fix: display records
The integration had the side effect of not
displaying records properly if it was enabled.

This is because the control flow would break out
of the loop that generates the records when
finding a LibKey link.

This bug mainly affected EBSCO host records.

This patch amends this so that records are
always displayed properly.

To replicate:

- checkout to a 25.01.00-based branch
- enable the EBSCO host module
- create an EBSCO host setting
- run a search in 'Article and Databases'
- notice that EBSCO host search results list
appears empty

Test plan:

- apply this patch
- run a search in 'Article and Databases'
- notice that EBSCO host search results display