DIS-138 fix: retraction notices for EBSCO EDS
This ensures that, should a retraction notice link
be sent back by LibKey for a given record, the
retraction link is used to populate the relevant
button, and the warning displayed to the user.

Before this, the url would have been overwritten,
and the 'retracted' variable set to true even if
a retraction was found.

For better readability, further amendments are
made to the control flow with a switch from nested
if statements to guard clauses.

Test plan

- run a search in 'Article and Databases'
- notice the 'Access Online' link to the right
of 'More Info' on each result
- check that the link opens the selected resource
in a new tab (this should take you through LibKey
and straight to the PDF document (if a pdf is
available for the resource))
- find a retracted article, notice that a
'Retracted' warning message appears, that it is a
link leading to LibKey page with information on
the retraction, and that the 'Access Online'
button is not available.