DIS-226: fix save URL to link image to
On the /WebBuilder/PortalRows page, the
'URL to link image to' cannot be saved.

To resolve this an $imageURL property needs to be
added to the PortalCell class, as this exists as
a field in the web_builder_portal_cell table,
which is where the value to assign to the URL to
link image to is pulled from.

Test plan: see Jira ticket DIS-194 and follow the
steps 1 - 7 in the test plan.

following this:

8. Image action choice will be preserved
but the 'URL to link image to' field will not show
immediately upon page reload
9. Upon selecting Image action again, 'URL to
link image to' will contain the previously entered
value.
10. On the Custom Web Builder Pages, select 'View'
11. Notice that clicking on the image will
redirect the user to the 'URL to link image to'.