DIS-98 refactor: check for ils consent support
This changes how Aspen determines whether the ILS it is set to use
support an ILS Consent integration, by implementing an
'hasIlsConsentSupport' method, instead of relying on 'method_exists'
checks.

The functionality, which remains unchanged, is as follows:

As of now, Koha is the only ILS for which this exists, and therefore,
for any other ILS, Administrators will not see the 'Enable ILS-issued
consents' setting option in Library Systems > Data Protection
Regulations.