DIS-222: Add notification to evergreen exporter if offline mode is enabled
notifying and exiting in evergreen exporter when offline mode is on.
Also noticed an issue where in offline mode the message noting that
the evergreen account profile could not be found would not occur if
encountered on the first loop because logEntry is never set. Set a
fallback indexingProfile variable so that message could get logged.