DIS-98 feat: improve info displayed to patron
Note:
Assumes that Koha has been set up as the ILS, and
that the koha-plugin-newsletter-consent plugin
is enabled.

Test plan:

- login or masquerade as a patron
- view the Privacy Settings page
- notice that information irrelevant to patron has
been removed from the tooltip
- notice that the formulation is improved ('to
receiving our Newsletter' instead of
'to our Newsletter')
- disable the Koha newsletter plugin
(or uninstall it)
- notice that the message showing on Privacy
Settings now shows on a blue background (instead
of red), and has been reworded for clarity.